requirements.txt
tensorflow-gpu==2.2.0
matplotlib==3.4.2
sklearn==0.0
opencv-python==4.5.2.54
mediapipe==0.8.6
pandas==1.1.3
keras==2.3.1
notebook==6.4.0
h5py==2.10.0
